[FEAT] category검색 메서드 추가

+++ src/main/java/org/appjam/bongbaek/domain/content/service/ContentService.java
@@ -1,8 +1,10 @@
 package org.appjam.bongbaek.domain.content.service;
 
 import lombok.RequiredArgsConstructor;
-import org.appjam.bongbaek.domain.content.dto.ContentWriteDto;
-import org.appjam.bongbaek.domain.content.dto.ContentDetailResponseDto;
+import org.appjam.bongbaek.domain.content.dto.response.ContentHomeResponseDto;
+import org.appjam.bongbaek.domain.content.dto.response.ContentListDto;
+import org.appjam.bongbaek.domain.content.dto.request.ContentWriteDto;
+import org.appjam.bongbaek.domain.content.dto.response.ContentDetailResponseDto;
 import org.appjam.bongbaek.domain.content.entity.Content;
 import org.appjam.bongbaek.domain.content.entity.ContentImage;
 import org.appjam.bongbaek.domain.content.repository.ContentRepository;
@@ -11,13 +13,20 @@ import org.appjam.bongbaek.global.exception.common.RequestInvalidException;
 import org.appjam.bongbaek.global.exception.content.ContentNotFoundException;
 import org.appjam.bongbaek.global.exception.image.ImageNotFoundException;
 import org.appjam.bongbaek.global.s3.dto.FileDto;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.util.List;
+
 @Service
 @RequiredArgsConstructor
 public class ContentService {
+    private final int PAGE_SIZE = 5;
+
     private final ContentRepository contentRepository;
     private final ContentImageService contentImageService;
 
@@ -61,7 +70,6 @@ public class ContentService {
         Content content = contentRepository.findById(contentId)
                 .orElseThrow(ContentNotFoundException::new);
 
-        String oldS3Key = content.getContentImage().getStorageKey();
         FileDto newMainImageDto = contentImageService.uploadImage(mainImageFile);
 
         ContentImage newMainImage = ContentImage.builder()
@@ -70,7 +78,11 @@ public class ContentService {
                 .build();
 
         content.uploadContentImage(newMainImage);
-        contentImageService.deleteImage(oldS3Key);
+
+        if (content.getContentImage() != null) {
+            String oldS3Key = content.getContentImage().getStorageKey();
+            contentImageService.deleteImage(oldS3Key);
+        }
     }
 
     @Transactional
@@ -93,5 +105,32 @@ public class ContentService {
                 .orElseThrow(ContentNotFoundException::new);
 
         return ContentDetailResponseDto.from(content);
+    }
+
+    @Transactional(readOnly = true)
+    public ContentHomeResponseDto getContentForHome() {
+        List<Content> contents = contentRepository.findTop3ByOrderByCreatedDateTimeDesc();
+
+        return ContentHomeResponseDto.from(contents);
+    }
+
+    @Transactional(readOnly = true)
+    public ContentListDto getContentList(int page, String category) {
+
+        Pageable pageable = PageRequest.of(page, PAGE_SIZE);
+        Category contentCategory = Category.of(category).orElse(null);
+
+        if(contentCategory == null) {
+            Page<Content> contents = contentRepository.findAllByOrderByCreatedDateTimeDesc(pageable);
+
+            return ContentListDto.of(contents);
+        }
+
+        Page<Content> contents = contentRepository.findContentsByContentCategoryOrderByCreatedDateTimeDesc(
+                Category.of(category).orElse(null),
+                pageable
+        );
+
+        return ContentListDto.of(contents);
     }
 }
